combined MapView into RestaurantPage
